requirements.txt
passlib
argon2_cffi
argon2pure
flask
flask_sqlalchemy
uwsgi